refactored code to seperate kubernetes resources
